Move delegators to avoid cyclic imports
